qmake: Introduce QMAKE_SUFFIX_STATICSHLIB for Unix makefile generator.
This is used to define the suffix of a static library in DLL templates.
This library is optional, it's built by explicitly giving the `staticlib`
target to the make command. A non-empty suffix is needed on e.g. OS/2
where this library's name clashes with the name of the import library.

Needed for #18.