qmake: Add support for TARGET_SHORT.
It is used in dynamic library templates where it's necessary to specify a different name for the DLL file (but keep the original one for the import library name and other related files) due to name length constraints. To be used on OS/2 for DLL targets when the original name doesn't fit 8 characters.

Closes #18.