corelib: Initialize QThreadPrivate vars on OS/2.
A lack of this could lead to e.g. terminatePending being interpreted as true at thread startup in setTerminationEnabled() which would result in a thread function hang before even entering QThread::run. Many tests would simply hang at the very beginning when attempting to start a WatchDog thread because of that.

See #29.