mkspecs: Add `map` CONFIG feature support on OS/2.
Adding `map` to CONFIG will cause the linker to generate `.map` files for each EXE and DLL target it builds. As opposed to Qt4, these files are not moved to the dist directory where the DLL/EXE goes, they are left out where the makefile is. This is done so that it usually makes no sense to distribute such files and they would only bloat the lib directory. Earlier we used them to create `.sym` files (which were definitely subject for distribution) but now we don't use `.sym` files as we switched to HLL debug info. This debug info is always left in-place in the respective EXE or DLL (no matter if it's a release build or a debug one) and can be later extracted into a separate `.dbg` file (this is done either manually with `emxomfstrip` or automatically by the RPM build framework when building an RPM).