tests: tst_largefile: Unmap before file removal.
On OS/2, a mapped region of a file holds a file descriptor so the file cannot be deleted until it is unmapped (so that the descriptor gets closed). I believe the very same problem whould have existed on Windows as well but this sub-test (mapOffsetOverflow) is disabled there.

Needed for #29.