corelib: Fix QDir::rmpath misbehavior revealed on OS/2.
The QFileSystemEngine::removeDirectory first called QDir::cleanPath and then searched for QDir::separator() when splitting up in separate leaves to remove. This is totally wrong as QDir::cleanPath guarantees `/` in the normalized path and QDir::separator() is `\` on OS/2. As a result, empty parent directories would not get removed despite the method semantics. This caused tst_QDir to fail in a couple of places on the second run due to non-removed test directories from the previous run.

This is an upstream Qt bug.

Needed for #29.