tests: tst_QDir: Align to OS/2.
Note that we disable some tests that require a network resource in the canonical form (\\server\share) when this resource is not available. It's not easy to make this work on OS/2 as this requires a classic LAN server instance which I don't have around (and this isn't so important nowadays).

Needed for #29.