tests: tst_QFileInfo: Align to OS/2.
This also fixes the isHidden test where the file existence wasn't checked so that isHidden would return false not only when the test file has no hidden attribute but also when it doesn't exist at all (QTBUG).

Note that we disable some tests that require a network resource in the canonical form (\\server\share) when this resource is not available. It's not easy to make this work on OS/2 as this requires a classic LAN server instance which I don't have around (and this isn't so important nowadays).

Needed for #29.