corelib: Fix recursion and multiple activation of zero timers.
Recursion and multiple activation within the same
QEventLoop::processEvents() call are both not allowed and there are
tests in tst_QTimer (see #29) that ensure this.

This commit also fixes starting/stopping an auxiliary thread to serve
non-zero timers (which is a 0a25375056b7bcdfd8d5ba9d4b0128cee7709366
regression).

Closes #52.