corelib: Restore qAppFileName on OS/2.
This is needed for QProcess. Using QCoreApplication::applicationFilePath
is not correct as it requires a QCoreApplication instance which is not
required by QProcess itself. While we could still make it work on OS/2
but it's not safe too because the value returned by this method may be
changed by QCoreApplicationPrivate::setApplicationFilePath to an
arbitrary value while QProcess needs a path to the real EXE image.