corelib/os2: Reimplement timer/socket thread machinery.
The new implementation has the following benefits:

- No constant thread start/stop on each processEvents call. The thread
is now kept running as long as there are socket notifiers and timers
that need it.

- Timer and socket events are now also processed correctly when the
application enters a nested PM message loop (e.g. when resizing or
moving windows using the mouse and standard frame controls). They
should work correctly even if the Qt event loop is not run at all
(i.e. if it's a plugin running in a native PM application and such).

This commit also adds QPMObjectWindow private helper class.

Closes #83.