fontdatabases: Support embedded bitmaps in fontconfig database.
For some unknown reason, the embedded bitmap setting from fontconfig
was completely ignored and fonts would be allowed to use such bitmaps
only if they are not antialased. This makes no sense as there may be
situations when the user wants the font both to use bitmaps at certain
sizes and be antialiased at all other sizes.

One such example is "Workplace Sans" which is a TTF that mimics the look
of the bitmap font "WarpSans" which is the system font on OS/2 Warp.

Needed for #85.