corelib: Avoid crashes when looking for Qt METADATA in plugin DLLs.
It is possible that the DLL is corrupt (or packed) so that the length
of the METADATA block is invalid and reading it would address space
beyond the file end causing an application crash.

This commit adds boundary checking and simply refuses to load the plugin
if the length value is benynd the file end.

Needed for #97.