qmake/os2: Add console/windows CONFIG option support.
Note that if neither `console` nor `windows` is explicitly specified in
CONFIG, then `windows` will be used if the target is a release build of
an application that uses QT's gui or widgets module. Otheriwse `console`
will be applied.

Note that Qt forces `console` for some of its parts (e.g. Qt tools
and all tests) which is perfectly fine on OS/2 as well.

Closes #99.