platforms/os2: Don't let titlebar disappear for default window position.
If the top-level window's position is not set explicitly (i.e. it's 0,0)
and the window has a frame, showing it on the screen would position
the client area at 0,0 making the title bar disappear beyond the top
edge of the screen. The Windows platform treats such a case as the
default one and moves the window so that its frame's position is at
0,0 instead. This commit implements the same for OS/2.

One of the affected real-life exapmles is Qt Linguist. See
https://github.com/bitwiseworks/qttools-os2/issues/1 for details.