corelib: Fix rounding to millisecond in QTimer.
This appears to be a QTBUG.

The millisecond resolution is used when detecting how long an event
dispatcher should wait for the next timer on Unix and OS/2. The
previous rounding code would incorrectly round values like .123000000
(i.e. with zero micro- and nanosecond parts) to .124000000 instead of
leaving them as is as they arlready have the right resolution.

This, in turn, broke tst_QTimer::remainingTimeInitial on platforms
where timers have the millisecond resolution (e.g. OS/2). It could
break other platforms too, it was just very unlikely that a timer
sample with zero micro- and nanosecond would be read at a test run
time.

This commit also fixes the QTIMERINFO_DEBUG code path (presumably
broken upstream when replacing timeval with timespec).

Needed for #110.