platforms/os2: Recognize Qt windows when changing focus.
Previously, if a Qt window got activated while some other Qt window
was active, the Qt focus window would be temporarily set to nullptr
before setting it to the new active window during this single focus
transition. This is not really necessary and the direct transition
is also asserted by tst_qguiapplication.cpp.

Needed for #70.