mkspecs/os2-g++: Don't force debug_and_release any more.
For RPM builds we need to disable it completely because the RPM devel
packages do not provide debug libraries and it simply makes no sense
to create two Makefile flavors: the debug one will never work.

Note that this commit also makes configure automatically append
debug_and_release to CONFIG in qconfig.pri when the feature is
auto-detected or explicitly enabled. Without this change it would
have to be forced in order to take effect even when building Qt
itself (looks like an upstream bug).